refactor: Lecture 내부 컨텐츠 컴포넌트로 분리
